Moved test fixes to test_suite_20
