+++ tests/sqlalchemy_test_suite/test_suite.py
@@ -7,9 +7,6 @@ from sqlalchemy.schema import Column, Sequence, Table
 from sqlalchemy.testing import config
 from sqlalchemy.testing.assertions import eq_
 from sqlalchemy.testing.suite import (
-    BizarroCharacterFKResolutionTest as _BizarroCharacterFKResolutionTest,
-)
-from sqlalchemy.testing.suite import (
     CompositeKeyReflectionTest as _CompositeKeyReflectionTest,
 )
 from sqlalchemy.testing.suite import FetchLimitOffsetTest as _FetchLimitOffsetTest
@@ -156,18 +153,3 @@ class CompositeKeyReflectionTest(_CompositeKeyReflectionTest):
     def test_pk_column_order(self):
         # Check https://snowflakecomputing.atlassian.net/browse/SNOW-640134 for details on breaking changes discussion.
         super().test_pk_column_order()
-
-
-class BizarroCharacterFKResolutionTest(_BizarroCharacterFKResolutionTest):
-    @testing.combinations(
-        ("id",), ("(3)",), ("col%p",), ("[brack]",), argnames="columnname"
-    )
-    @testing.variation("use_composite", [True, False])
-    @testing.combinations(
-        ("plain",),
-        ("(2)",),
-        ("[brackets]",),
-        argnames="tablename",
-    )
-    def test_fk_ref(self, connection, metadata, use_composite, tablename, columnname):
-        super().test_fk_ref(connection, metadata, use_composite, tablename, columnname)
