+++ tests/sqlalchemy_test_suite/test_suite_20.py
@@ -75,6 +75,10 @@ class InsertBehaviorTest(_InsertBehaviorTest):
     def test_empty_insert_multiple(self, connection):
         pass
 
+    @pytest.mark.skip("Snowflake does not support returning in insert.")
+    def test_no_results_for_non_returning_insert(self, connection, style, executemany):
+        pass
+
 
 # road to 2.0
 class TrueDivTest(_TrueDivTest):
@@ -88,10 +92,6 @@ class TimeMicrosecondsTest(_TimeMicrosecondsTest):
 
 
 class DateTimeHistoricTest(_DateTimeHistoricTest):
-    ...
-
-
-class BizarroCharacterFKResolutionTest(_BizarroCharacterFKResolutionTest):
     ...
 
 
@@ -174,3 +174,18 @@ class CompositeKeyReflectionTest(_CompositeKeyReflectionTest):
     def test_pk_column_order(self):
         # Check https://snowflakecomputing.atlassian.net/browse/SNOW-640134 for details on breaking changes discussion.
         super().test_pk_column_order()
+
+
+class BizarroCharacterFKResolutionTest(_BizarroCharacterFKResolutionTest):
+    @testing.combinations(
+        ("id",), ("(3)",), ("col%p",), ("[brack]",), argnames="columnname"
+    )
+    @testing.variation("use_composite", [True, False])
+    @testing.combinations(
+        ("plain",),
+        ("(2)",),
+        ("[brackets]",),
+        argnames="tablename",
+    )
+    def test_fk_ref(self, connection, metadata, use_composite, tablename, columnname):
+        super().test_fk_ref(connection, metadata, use_composite, tablename, columnname)
